Update EthTx commands to use the TableRenderer interface
